backend authorization on userprofile, comment, tag, and category

+++ Tabloid/Controllers/TagController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using Tabloid.Models;
 using Tabloid.Repositories;
+using System.Security.Claims;
 
 namespace Tabloid.Controllers
 {
@@ -12,9 +13,13 @@ namespace Tabloid.Controllers
     public class TagController : ControllerBase
     {
         private readonly ITagRepository _tagRepository;
-        public TagController(ITagRepository tagRepository)
+        private readonly IUserProfileRepository _userProfileRepository;
+
+        public TagController(ITagRepository tagRepository,
+            IUserProfileRepository userProfileRepository)
         {
             _tagRepository = tagRepository;
+            _userProfileRepository = userProfileRepository;
         }
         [HttpGet]
         public IActionResult Get()
@@ -37,6 +42,11 @@ namespace Tabloid.Controllers
 
         public IActionResult Post(Tag tag)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+            if (currentUserProfile.UserTypeId != 1)
+            {
+                return Unauthorized();
+            }
             _tagRepository.AddTag(tag);
             return CreatedAtAction("Get", new { id = tag.Id }, tag);
         }
@@ -45,6 +55,11 @@ namespace Tabloid.Controllers
 
         public IActionResult Put(int Id, Tag tag)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+            if (currentUserProfile.UserTypeId != 1)
+            {
+                return Unauthorized();
+            }
             if (tag.Id != Id)
             {
                 return BadRequest();
@@ -57,8 +72,18 @@ namespace Tabloid.Controllers
         [HttpDelete("{Id}")]
         public IActionResult Delete(int Id)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+            if (currentUserProfile.UserTypeId != 1)
+            {
+                return Unauthorized();
+            }
             _tagRepository.Delete(Id);
             return NoContent();
+        }
+        private UserProfile GetCurrentUserProfile()
+        {
+            var firebaseUserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            return _userProfileRepository.GetByFirebaseUserId(firebaseUserId);
         }
     }
 }
